genre modified

+++ routes/genres.mjs
@@ -7,14 +7,16 @@ import { Author } from "../mongoose/schemas/author.mjs"
 import { Genre } from "../mongoose/schemas/genre.mjs"
 import { Publisher } from "../mongoose/schemas/publisher.mjs"
 import bodyParser from "body-parser"
+import mongoose from "mongoose"
 
 
 const router = Router()
 
 //add new genre
 router.post("/api/genres", adminAuth, async (request, response) => {
-    const {genre} = request.body
+    
     try {
+        const {genre} = request.body
         const myGenre = await Genre.create({genre})
         return response.status(201).send(myGenre)
     } catch (error){
@@ -23,10 +25,16 @@ router.post("/api/genres", adminAuth, async (request, response) => {
 })
 
 //delete a genre
-router.delete("/api/genres", adminAuth, async (request, response) => {
-    const {_id} = request.body
+router.delete("/api/genres/:_id", adminAuth, async (request, response) => {
+    
     try {
+        const {_id} = request.params
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return response.status(400).send({ message: 'Invalid ID format' });
+            }
+
         const deletedGenre = await Genre.findByIdAndDelete(_id)
+
         if (deletedGenre){
             return response.status(201).send(`Deleted genre: ${deletedGenre}`)
         } else {
@@ -38,9 +46,13 @@ router.delete("/api/genres", adminAuth, async (request, response) => {
 })
 
 //get detail of an genre
-router.get("/api/genres/detail", adminAuth, async (request, response) => {
-    const {_id} = request.body
+router.get("/api/genres/:_id", adminAuth, async (request, response) => {
+   
     try {
+        const {_id} = request.params
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return response.status(400).send({ message: 'Invalid ID format' });
+            }
         const queriedGenre = await Genre.findById(_id)
         if (queriedGenre){
             return response.status(201).send(queriedGenre)
@@ -53,14 +65,12 @@ router.get("/api/genres/detail", adminAuth, async (request, response) => {
 } )
 
 //list genre
-router.get("/api/genres/list", adminAuth, async (request, response) => {
+router.get("/api/genres", adminAuth, async (request, response) => {
     try {
         const genreList = await Genre.find().sort({ createdAt : -1 })
-        if (genreList.length > 0){
-            return response.status(201).send(genreList)
-        } else {
-            return response.send("The list is empty")
-        }
+      
+        return response.status(201).send(genreList)
+       
     } catch (error) {
         return response.status(400).send(error.message)
     }
@@ -68,33 +78,28 @@ router.get("/api/genres/list", adminAuth, async (request, response) => {
 
 
 //update genre
-router.patch("/api/genres", adminAuth, async (request, response) => {
-    let isError = []
-    const {_id} = request.body
-    if (!_id){
-        return response.status(400).send(`id is missing`)
-    }
+router.patch("/api/genres/:_id", adminAuth, async (request, response) => {
+    try {
+    const {_id} = request.params
+    if (!mongoose.Types.ObjectId.isValid(_id)) {
+        return response.status(400).send({ message: 'Invalid ID format' });
+        }
     const myGenre = await Genre.findById(_id)
-
+    if (!myGenre){
+        return response.status(400).send({ message: 'Cannot query document' });
+    }
     Object.entries(request.body).forEach(([key, value]) => {
         if (value){
           myGenre[key] = value
-        } else {
-          isError.push(key)
         }
       })
 
-    if (isError.length >= 1){
-        return response.status(400).send(`You are not allowed update "${isError}"`)  
-      }
-        
-    try {
+    
         const updatedGenre = await myGenre.save()
         return response.status(201).send(updatedGenre)
         }
     catch (error) {  
-        console.log(error)
-        return response.sendStatus(400)
+        return response.status(400).send(error.message)
         }
 })
